T: Split old plain hints' tests into "parameter" and "type"
